sql isolation
The beta site is now isolated from the main site in that it has its own
database. It will no longer get updates to modules or calanders from the
main site and changes with modules location and content will not get
carried over to main site. This has been tested and confirmed. Forums
are still and will stay being shared cross-site.